// mitm_logs/arda-api-sdk-browser/api.ts
export interface ApiAdminShopOrderOrderIdPutRequestOneOf {
    /**
     * 
     * @type {string}
     * @memberof ApiAdminShopOrderOrderIdPutRequestOneOf
     */
    'action': ApiAdminShopOrderOrderIdPutRequestOneOfActionEnum;
    /**
     * 
     * @type {ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHold}
     * @memberof ApiAdminShopOrderOrderIdPutRequestOneOf
     */
    'shipmentHold': ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHold;
}

export const ApiAdminShopOrderOrderIdPutRequestOneOfActionEnum = {
    AdminAddShippingHold: 'AdminAddShippingHold'
} as const;

export type ApiAdminShopOrderOrderIdPutRequestOneOfActionEnum = typeof ApiAdminShopOrderOrderIdPutRequestOneOfActionEnum[keyof typeof ApiAdminShopOrderOrderIdPutRequestOneOfActionEnum];

/**
 * 
 * @export
 * @interface ApiAdminShopOrderOrderIdPutRequestOneOf1
 */
export interface ApiAdminShopOrderOrderIdPutRequestOneOf1 {
    /**
     * 
     * @type {string}
     * @memberof ApiAdminShopOrderOrderIdPutRequestOneOf1
     */
    'action': ApiAdminShopOrderOrderIdPutRequestOneOf1ActionEnum;
}

export const ApiAdminShopOrderOrderIdPutRequestOneOf1ActionEnum = {
    AdminRemoveShippingHold: 'AdminRemoveShippingHold'
} as const;

export type ApiAdminShopOrderOrderIdPutRequestOneOf1ActionEnum = typeof ApiAdminShopOrderOrderIdPutRequestOneOf1ActionEnum[keyof typeof ApiAdminShopOrderOrderIdPutRequestOneOf1ActionEnum];
// ...
export interface ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHold {
    /**
     * 
     * @type {string}
     * @memberof ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHold
     */
    'reason': ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHoldReasonEnum;
    /**
     * 
     * @type {string}
     * @memberof ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHold
     */
    'notes'?: string;
}

export const ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHoldReasonEnum = {
    AddressIncorrect: 'AddressIncorrect',
    InventoryUnavailable: 'InventoryUnavailable',
    WaitingOnReturn: 'WaitingOnReturn',
    FraudRisk: 'FraudRisk',
    WaitingOnCustomerDecision: 'WaitingOnCustomerDecision',
    Other: 'Other'
} as const;

export type ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHoldReasonEnum = typeof ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHoldReasonEnum[keyof typeof ApiAdminShopOrderOrderIdPutRequestOneOfShipmentHoldReasonEnum];